Make class Recipe public in separate .java file
